Split out config module to crate
