+++ crates/genesis_config/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "genesis_config"
+authors = ["Ben Askew-Gale <bengale2007@googlemail.com>"]
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+
+[dependencies]
+bevy = { workspace = true }
+once_cell = { workspace = true}
+serde_derive = { workspace = true}
+confy = { workspace = true}
+serde = { workspace = true}
