+++ crates/genesis_lib/Cargo.toml
@@ -12,14 +12,13 @@ genesis_newtype = { path = "../genesis_newtype/"}
 genesis_color = { path = "../genesis_color/"}
 genesis_maths = {path = "../genesis_maths/"}
 genesis_spawners = { path = "../genesis_spawners" }
+genesis_config = { path = "../genesis_config" }
 bevy = { workspace = true }
 rand = { workspace = true }
 thiserror = { workspace = true }
 derive_more = { workspace = true }
 bevy_rapier2d = { workspace = true }
 anyhow = { workspace = true }
-once_cell = { workspace = true}
-confy = { workspace = true}
 serde_derive = { workspace = true}
 serde = { workspace = true}
 bevy_egui = { workspace = true}
